Add an explicit cast from uint64_t to size_t.
Otherwise it fails to compile on Chromium 32-bit platforms.

gfe-relnote: n/a (no functional change)
PiperOrigin-RevId: 229406761

+++ http2/hpack/decoder/hpack_string_decoder.h
@@ -168,7 +168,8 @@ class HTTP2_EXPORT_PRIVATE HpackStringDecoder {
   // false otherwise, in which case status set.
   template <class Listener>
   void OnStringStart(Listener* cb, DecodeStatus* status) {
-    remaining_ = length_decoder_.value();
+    // TODO(vasilvv): fail explicitly in case of truncation.
+    remaining_ = static_cast<size_t>(length_decoder_.value());
     // Make callback so consumer knows what is coming.
     cb->OnStringStart(huffman_encoded_, remaining_);
   }
